update code after testing on mac locally

+++ engine_app/monitor.py
@@ -4,25 +4,28 @@ import threading
 import requests
 import json
 import logging
-from .modules.config import load_config
-from .modules.sftp_utils import connect_sftp, move_folder
-from .modules.logging_config import configure_logging
-from .modules.folder_processor import is_new_subfolder, generate_random_filename
+import stat
+from modules.config import load_config
+from modules.sftp_utils import connect_sftp, move_folder, check_folder_exists
+from modules.logging_config import configure_logging
+from modules.folder_processor import is_new_subfolder, generate_random_filename
 
 configure_logging('logs/monitor.log')
 
 def check_for_new_folders(sftp, monitored_folder, known_folders):
     """Check for new folders in the monitored folder."""
     try:
-        current_folders = sftp.listdir(monitored_folder)
+        # List items in the monitored folder
+        current_items = sftp.listdir_attr(monitored_folder)
+        # Filter out files and only include directories
+        current_folders = [
+            item.filename for item in current_items if stat.S_ISDIR(item.st_mode)
+        ]
         new_folders = [folder for folder in current_folders if folder not in known_folders]
         return new_folders
     except Exception as e:
         logging.error(f"Error checking new folders: {e}")
         return []
-import requests
-import json
-import logging
 
 def execute_requests_command(folder, engine_url):
     """Execute a POST request to process the folder and return the processed field value."""
@@ -100,24 +103,26 @@ def target_new_folder(sftp, sftp_config, folder, engine_url):
 
     if not processed:
         with connect_sftp(sftp_config) as sftp_move:
-            logging.error(f"Folder processing failed after {max_retries} attempts: {folder}. Moving to failed category")   
-            random_filename = generate_random_filename()
-            local_results_path = os.path.join("/tmp", random_filename)
-            error = {"Error": "Processing error for folder : Engine call from monitor failed"}
+            logging.error(f"Folder processing failed after {max_retries} attempts: {folder}. Moving to failed category")
+            if check_folder_exists(sftp_move, folder, sftp_config['SFTP_FOLDER']):
+                logging.error(f" Moving to failed category")  
+                random_filename = generate_random_filename()
+                local_results_path = os.path.join("/tmp", random_filename)
+                error = {"Error": "Processing error for folder : Engine call from monitor failed"}
 
-            with open(local_results_path, 'w') as f:
-                json.dump(error, f)
-            logging.info(f"Created temporary local file with random name: {local_results_path}")
+                with open(local_results_path, 'w') as f:
+                    json.dump(error, f)
+                logging.info(f"Created temporary local file with random name: {local_results_path}")
 
-            result_path = os.path.join(sftp_config['SFTP_FOLDER'], folder, 'error.json')
-            logging.info(f"result_path {result_path},{local_results_path} ")
-            try:
-                sftp_move.put(local_results_path, result_path)
-            except FileNotFoundError:
-                logging.warning(f"File not found during upload: {local_results_path}")
-            logging.info(f"Successfully uploaded the error results as error.json for subfolder {folder}")
-            os.remove(local_results_path)
-            move_folder(sftp_move, sftp_config['SFTP_FOLDER'], sftp_config['FAILED_SFTP_FOLDER'], folder)
+                result_path = os.path.join(sftp_config['SFTP_FOLDER'], folder, 'error.json')
+                logging.info(f"result_path {result_path},{local_results_path} ")
+                try:
+                    sftp_move.put(local_results_path, result_path)
+                except FileNotFoundError:
+                    logging.warning(f"File not found during upload: {local_results_path}")
+                logging.info(f"Successfully uploaded the error results as error.json for subfolder {folder}")
+                os.remove(local_results_path)
+                move_folder(sftp_move, sftp_config['SFTP_FOLDER'], sftp_config['FAILED_SFTP_FOLDER'], folder)
 
     # Safely remove the folder from known_folders
